+++ tests/test_geometric_tests.py
@@ -202,6 +202,15 @@ class test_geometric_tests(unittest.TestCase):
         self.assertEqual(result, None)
 
     def test_point_height_above_plane(self):
+        pl = plane.create([0., 1., 0.], -1.)
+        p = np.array([0., 1., 0.])
+        result = gt.point_height_above_plane(p, pl)
+        self.assertEqual(result, 0.)
+
+        p = np.array([0., 0., 0.])
+        result = gt.point_height_above_plane(p, pl)
+        self.assertEqual(result, -1.)
+
         v1 = np.array([ 0.0, 0.0, 1.0])
         v2 = np.array([ 1.0, 0.0, 1.0])
         v3 = np.array([ 0.0, 1.0, 1.0])
@@ -211,9 +220,9 @@ class test_geometric_tests(unittest.TestCase):
         pl = plane.invert_normal(pl)
 
         result = gt.point_height_above_plane(p, pl)
-        self.assertEqual(result, -19.)
+        self.assertEqual(result, 19.)
 
-        pl = plane.create_xz(distance=5.)
+        pl = plane.create_xz(distance=-5.)
         p = np.array([0., 5., 0.])
         h = gt.point_height_above_plane(p, pl)
         self.assertEqual(h, 0.)